+++ src/ska_tmc_cdm/messages/central_node/release_resources.py
@@ -44,7 +44,7 @@ class ReleaseResourcesRequest(CdmObject):
 
     interface: Optional[str] = SCHEMA
     transaction_id: Optional[str] = None
-    subarray_id: Optional[int] = None
+    subarray_id: Optional[int] = Field(default=None, ge=1, le=16)
     release_all: StrictBool = False
     dish: Optional[DishAllocation] = Field(
         default=None,
@@ -54,7 +54,7 @@ class ReleaseResourcesRequest(CdmObject):
         ),
     )
 
-    # Custom logic required to mimic the behavior of marshallow.fields.Pluck()
+    # Custom logic required to mimic the behaviour of marshallow.fields.Pluck()
     # to maintain backwards compatibility when removing Marshmallow schemas.
     # https://marshmallow.readthedocs.io/en/stable/marshmallow.fields.html#marshmallow.fields.Pluck
     @field_serializer("dish", when_used="json-unless-none")